+++ src/tran.c
@@ -472,6 +472,12 @@ int new_or_ith(const char *s) {
     tb.ix=-2; 
     return 0;
   }
+  if (!strcmp("ifelse", s)){
+    updateSyntaxCol();
+    trans_syntax_error_report_fn(_("'ifelse' cannot be a variable/state in an RxODE model"));
+    tb.ix=-2;
+    return 0;
+  }
   if (!strcmp("ss", s)){
     updateSyntaxCol();
     trans_syntax_error_report_fn(_("'ss' cannot be a variable in an RxODE model"));